+++ samples/star/src/commonMain/kotlin/com/slack/circuit/star/petdetail/PetDetailScreen.kt
@@ -73,6 +73,7 @@ data class PetDetailScreen(val petId: Long, val photoUrlMemoryCacheKey: String?)
     data object UnknownAnimal : State
 
     data class Success(
+      val animal: Animal,
       val url: String,
       val photoUrls: ImmutableList<String>,
       val photoUrlMemoryCacheKey: String?,
@@ -94,6 +95,7 @@ internal fun Animal.toPetDetailState(
   eventSink: (Event) -> Unit,
 ): State {
   return Success(
+    animal = this,
     url = url,
     photoUrls = photoUrls,
     photoUrlMemoryCacheKey = photoUrlMemoryCacheKey,
@@ -197,6 +199,7 @@ private fun ShowAnimal(state: Success, padding: PaddingValues) {
     movableContentOf {
       CircuitContent(
         PetPhotoCarouselScreen(
+          id = state.animal.id,
           name = state.name,
           photoUrls = state.photoUrls,
           photoUrlMemoryCacheKey = state.photoUrlMemoryCacheKey,
